23: moved helpers to their right place
