+++ app/cv/page.tsx
@@ -2,9 +2,9 @@
 import { CkH2 } from "@/components/ui/typography";
 import { ContactInformation } from "./contact-information/contact-infomation";
 import { cvData } from "./data";
-import { dataToSidebarContent } from "./helpers";
 import TextSection from "./shared/text-section";
 import { Sidebar } from "./sidebar/sidebar";
+import { dataToSidebarContent } from "./sidebar/sidebar-helpers";
 import SkillsSection from "./skills/skills-section";
 import WorkExperienceCard from "./work-experience/work-experience-card";
